+++ sites/partners/lib/helpers.ts
@@ -1,7 +1,9 @@
 import { SetStateAction } from "react"
 import { t, CloudinaryUpload, TimeFieldPeriod } from "@bloom-housing/ui-components"
 import { cloudinaryUrlFromId } from "@bloom-housing/shared-helpers"
-import moment from "moment"
+import dayjs from "dayjs"
+import utc from "dayjs/plugin/utc"
+dayjs.extend(utc)
 import {
   ApplicationSubmissionType,
   AssetsService,
@@ -9,7 +11,7 @@ import {
   ListingEvent,
   IncomePeriod,
 } from "@bloom-housing/backend-core/types"
-import { TempUnit, FormListing } from "../src/listings/PaperListingForm"
+import { TempUnit, FormListing } from "../src/listings/PaperListingForm/formTypes"
 import { FieldError } from "react-hook-form"
 
 type DateTimePST = {
@@ -74,10 +76,10 @@ export const convertDataToPst = (dateObj: Date, type: ApplicationSubmissionType)
   }
 
   if (type === ApplicationSubmissionType.paper) {
-    const momentDate = moment(dateObj)
+    const dayjsDate = dayjs(dateObj)
 
-    const date = momentDate.utc().format("MM/DD/YYYY")
-    const time = momentDate.utc().format("hh:mm:ss A")
+    const date = dayjsDate.utc().format("MM/DD/YYYY")
+    const time = dayjsDate.utc().format("hh:mm:ss A")
 
     return {
       date,